Implement streaming to and from comby for indexed structural search
This commit modifies the path for indexed structural search to use the "-tar" flag to stream candidate matches to comby's stdin in tar format. We also now can stream results from comby's stdout. This should significantly improve the performance of indexed structural search.

The path for unindexed structural search remains untouched. I _think_ it should be possible to stream from stdout for unindexed search as well but I haven't been able to get that working yet so setting a checkpoint here with things in a working state.