codemirror blob view: First pass on porting blob tests
This commit ports the non-skipped tests from `blob-viewer.test.ts` to
test them against the CodeMirror implementation. I had to make some
minor adjustments to make it work with CodeMirror's DOM structure.

The most significant change is probably how "selections" work. Instead
of simply clicking on a line (which did not work) I extended the
existing editor API to allow us to specify the position of the text
cursor.
That only works because currently the view is configured to show/have a
cursor. If we are not keeping this functionality we have to find another
approach.

A couple of features, especially extensions-related ones, don't seem to
be tested in the existing integration test. I plan to add more tests for
CodeMirror specifically in future PRs.